REF: Split up parsers into smaller files
