+++ pandas/io/json/_json.py
@@ -38,7 +38,7 @@ from pandas.io.common import (
 )
 from pandas.io.json._normalize import convert_to_line_delimits
 from pandas.io.json._table_schema import build_table_schema, parse_table_schema
-from pandas.io.parsers import validate_integer
+from pandas.io.parsers.readers import validate_integer
 
 loads = json.loads
 dumps = json.dumps
