+++ pandas/tests/io/parser/test_dialect.py
@@ -98,9 +98,9 @@ def test_dialect_conflict_except_delimiter(all_parsers, custom_dialect, arg, val
         if "value" == "dialect":  # No conflict --> no warning.
             kwds[arg] = dialect_kwargs[arg]
         elif "value" == "default":  # Default --> no warning.
-            from pandas.io.parsers import _parser_defaults
+            from pandas.io.parsers.base_parser import parser_defaults
 
-            kwds[arg] = _parser_defaults[arg]
+            kwds[arg] = parser_defaults[arg]
         else:  # Non-default + conflict with dialect --> warning.
             warning_klass = ParserWarning
             kwds[arg] = "blah"
