+++ pandas/tests/io/parser/test_read_fwf.py
@@ -11,11 +11,13 @@ from pathlib import Path
 import numpy as np
 import pytest
 
+from pandas.errors import EmptyDataError
+
 import pandas as pd
 from pandas import DataFrame, DatetimeIndex
 import pandas._testing as tm
 
-from pandas.io.parsers import EmptyDataError, read_csv, read_fwf
+from pandas.io.parsers.readers import read_csv, read_fwf
 
 
 def test_basic():
