+++ pandas/tests/io/parser/test_unsupported.py
@@ -14,8 +14,8 @@ from pandas.errors import ParserError
 
 import pandas._testing as tm
 
-import pandas.io.parsers as parsers
-from pandas.io.parsers import read_csv
+import pandas.io.parsers.readers as parsers
+from pandas.io.parsers.readers import read_csv
 
 
 @pytest.fixture(params=["python", "python-fwf"], ids=lambda val: val)
@@ -86,7 +86,7 @@ x   q   30      3    -0.6662 -0.5243 -0.3580  0.89145  2.5838"""
             read_csv(StringIO(data), lineterminator="~~")
 
     def test_python_engine(self, python_engine):
-        from pandas.io.parsers import _python_unsupported as py_unsupported
+        from pandas.io.parsers.readers import _python_unsupported as py_unsupported
 
         data = """1,2,3,,
 1,2,3,4,
